refactor(core): move the tour reducer out from under React (v2 §1.3a)
tourReducer, its four handlers, findAutoStartTour and MAX_HIDDEN_CHAIN move
verbatim from tour-provider.tsx into lib/tour-engine/reducer.ts. Pure code
motion — the reducer was already pure and already imported nothing from React;
the provider now imports it and keeps dispatching through useReducer.

The reducer was previously exercised only through provider integration tests.
reducer.test.ts covers every switch arm directly, including the two that carry
real logic: UPDATE_TOURS's shallow-identity fast path and
HYDRATE_TERMINAL_TOURS's union merge.

tour-provider.tsx: 1431 -> 1252 lines.
